Add new script to demonstrate map function application
A new script 'map.ts' has been added under the /examples directory. This script provides an illustration of how to use our pre-defined map function on an array of numbers with callback functions. It helps to depict the use of function as an argument and its application on every element in the array.